Applying patch from Joanne Tong (joannet@ca.ibm.com) for Bugzilla bug report 16311.  After completing the relevant operations in applyTemplateToNode, the method was popping various pieces of the context off of stacks.  However, in the case of the default template rule for text nodes, it was trying to pop variables off the stack, even though no variables had been pushed onto the stack.
git-svn-id: https://svn.eu.apache.org/repos/asf/xalan/java/trunk@337647 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/xalan/transformer/TransformerImpl.java
@@ -2192,7 +2192,8 @@ public class TransformerImpl extends Transformer
     }
     finally
     {
-      m_xcontext.getVarStack().unlink();
+      if (!isDefaultTextRule)
+        m_xcontext.getVarStack().unlink();
       m_xcontext.popCurrentNode();
       if (!isApplyImports) {
           m_xcontext.popContextNodeList();
